requirements.txt
flask
flask-cors
PyJWT
python-dotenv